Monitoring/requirements.txt
django
djangorestframework
drf_restwind
gunicorn
requests
httpx
celery[redis]
prometheus-client